Slowmode V2 update!
Commands to enable, disable and list channels with slowmode currently
enabled.
Slowmode deletes messages a user post if it has been less then 5 seconds
from their last posting.
If a user has "manage_messages" enabled on a channel, they are ignored.
Bot will also warn you in a channel that it no longer has the
"manage_messages" perm for every 10 messages.
Slowmode is currently set to reset the timer for every message sent,
even if the bot deletes it in slowmode.
If a user continues to try to send messages, after 3 messages the bot
will mention them with a warning the that channel is in slowmode and
that they must wait 5 seconds between posting.